Allow to load in json context so that references can be loaded in
This follows the best practice on how the jsonschema expects additional schemas to be loaded. One bug I did find is that the required argument is not fully supported for additional required json schemas. I'm trying to replicate this and create a bug report for the jsonschema module but to that time being, this fixes it for me.